Rename module 'carga' to 'workload'.
