+++ src/ispd/arquivo/xml/IconicoXML.java
@@ -11,10 +11,10 @@ import ispd.gui.PickModelTypeDialog;
 import ispd.gui.iconico.Edge;
 import ispd.gui.iconico.Vertex;
 import ispd.gui.iconico.grade.VirtualMachine;
-import ispd.motor.carga.impl.CollectionWorkloadGenerator;
-import ispd.motor.carga.impl.TraceFileWorkloadGenerator;
-import ispd.motor.carga.impl.GlobalWorkloadGenerator;
-import ispd.motor.carga.WorkloadGenerator;
+import ispd.motor.workload.impl.CollectionWorkloadGenerator;
+import ispd.motor.workload.impl.TraceFileWorkloadGenerator;
+import ispd.motor.workload.impl.GlobalWorkloadGenerator;
+import ispd.motor.workload.WorkloadGenerator;
 import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.RedeDeFilasCloud;
 import org.w3c.dom.Document;
