+++ src/ispd/arquivo/xml/models/builders/LoadBuilder.java
@@ -3,12 +3,12 @@ package ispd.arquivo.xml.models.builders;
 import ispd.arquivo.xml.utils.WrappedDocument;
 import ispd.arquivo.xml.utils.WrappedElement;
 import ispd.motor.random.TwoStageUniform;
-import ispd.motor.carga.impl.CollectionWorkloadGenerator;
-import ispd.motor.carga.impl.PerNodeWorkloadGenerator;
-import ispd.motor.carga.impl.GlobalWorkloadGenerator;
-import ispd.motor.carga.impl.TraceFileWorkloadGenerator;
-import ispd.motor.carga.WorkloadGenerator;
-import ispd.motor.carga.WorkloadGeneratorType;
+import ispd.motor.workload.impl.CollectionWorkloadGenerator;
+import ispd.motor.workload.impl.PerNodeWorkloadGenerator;
+import ispd.motor.workload.impl.GlobalWorkloadGenerator;
+import ispd.motor.workload.impl.TraceFileWorkloadGenerator;
+import ispd.motor.workload.WorkloadGenerator;
+import ispd.motor.workload.WorkloadGeneratorType;
 
 import java.io.File;
 import java.util.Optional;
