+++ src/ispd/gui/iconico/grade/DesenhoGrade.java
@@ -7,12 +7,12 @@ import ispd.gui.iconico.DrawingArea;
 import ispd.gui.iconico.Edge;
 import ispd.gui.iconico.Icon;
 import ispd.gui.iconico.Vertex;
-import ispd.motor.carga.impl.PerNodeWorkloadGenerator;
-import ispd.motor.carga.impl.CollectionWorkloadGenerator;
-import ispd.motor.carga.impl.GlobalWorkloadGenerator;
-import ispd.motor.carga.impl.TraceFileWorkloadGenerator;
-import ispd.motor.carga.WorkloadGenerator;
-import ispd.motor.carga.WorkloadGeneratorType;
+import ispd.motor.workload.impl.PerNodeWorkloadGenerator;
+import ispd.motor.workload.impl.CollectionWorkloadGenerator;
+import ispd.motor.workload.impl.GlobalWorkloadGenerator;
+import ispd.motor.workload.impl.TraceFileWorkloadGenerator;
+import ispd.motor.workload.WorkloadGenerator;
+import ispd.motor.workload.WorkloadGeneratorType;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
