+++ src/ispd/motor/workload/impl/GlobalWorkloadGenerator.java
@@ -1,6 +1,6 @@
-package ispd.motor.carga.impl;
+package ispd.motor.workload.impl;
 
-import ispd.motor.carga.WorkloadGeneratorType;
+import ispd.motor.workload.WorkloadGeneratorType;
 import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.Tarefa;
 import ispd.motor.filas.servidores.CS_Processamento;
