+++ src/ispd/motor/workload/task/ExternalTraceTaskBuilder.java
@@ -1,4 +1,4 @@
-package ispd.motor.carga.task;
+package ispd.motor.workload.task;
 
 import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.Tarefa;
